print_long function fixed
https://codereview.stackexchange.com/questions/255160/c-text-word-wrapping/255164

+++ src/util.h
@@ -31,4 +31,4 @@ struct tm *tm_empty(void);
 char *tm_tascii(const struct tm *);
 char *tm_dascii(const struct tm *, const struct config *);
 
-void print_long(const char *, int);
+void print_long(const char *, size_t);
